requirements.txt
Flask
configparser==3.5.0
python-telegram-bot
